requirements.txt
discord.py
ics
requests
datetime
python-dotenv
audioop-lts
Jinja2
imgkit